Add an example of cnn model
