+++ mlpipe/__init__.py
@@ -5,3 +5,6 @@
 __author__ = """Yilun Guan"""
 __email__ = 'zoom.aaron@gmail.com'
 __version__ = '0.1.0'
+
+
+from mlpipe import *
